Split the output tests class
Debugging the testcases in the class is quite heavy, so refactor it
into a abstract class, move the abstract class into __init__.py to be
reused from a class with a single input file.
